// Backend/src/domain/models/blog.ts
import {IUser} from './user'


export interface IBlog {
  _id?: string;                
  title: string;              
  content: string;
  category:string;           
  image: string;              
  userId?: IUser|string;               
  createdAt?: Date;           // Optional: auto-generated by Mongoose
  updatedAt?: Date;           // Optional: auto-generated by Mongoose
}